requirements.txt
tensorflow==1.15.
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
funcsigs==1.0.2
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
mock==2.0.0
networkx==2.1
numpy==1.14.2
opencv-python==3.4.0.12
pbr==4.0.0
Pillow==5.0.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.1
six==1.11.0
tensorflow-tensorboard==0.1.8
tqdm==4.19.9
Werkzeug==0.14.1
